Separating gui mainWindow from helper classes
